backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pydantic
aiofiles
razorpay
supabase
python-dotenv
httpx